Correctly remove orphaned css modules
With css during a live-reload we wait until the new `<style>` is
inserted into the page before removing the old `<style>`. The reason for
this is so that there is no flash of the css being temporarily missing.

So to make orphan removal work we need to use a flag that lets us know
if the css module is reloaded. it works like there:

1. A `wasReloaded` flag is initially false.
2. We listen for when the css module is reloaded.
3. If it is reloaded the `wasReloaded` flag gets set to true.
4. Meanwhile we are also listening for the css module's disposal. If it
is disposed it **might** be reloaded, this is why we need the flag, we
don't know if it will be reloaded or not.
5. We then listen to the cycleComplete event to know that all reloading
is complete, at which time we can check the flag. If the flag is still
`false` we know the module was disposed by never reloaded, and therefore
is an orphan that can be removed.

Fixes #7