Know when to add a style by making reloaded styles as dirty
During live-reload we need this order to happen:

1. New version of the css is appended to the head.
2. Old version of the css is removed from the head.

This prevents FOUC.

However live-reload works in this order:

1. Old modules are disposed.
2. New modules are imported.

As you can see, this is the opposite order from what we want with css.

This is essentially what makes css live-reload harder than for other
plugins.

To make the situation harder we have server side rendering where a
`<style>` will already be in th page.  In this case we want to not
insert a new style.

So, the question is how do we know whether to insert a new style or not?
If we are live-reloading we want to. If we are loading for the first
time with ssr enabled we don't want to.

The solution is to never insert a new `<style>` if an existing one is
there, **unless** that style is marked dirty. We mark it dirty when it
is being disposed.

Fixes #23