Refactors live-reload support to be simplified.
This change refactors our live-reload support to make it simplified.
Instead of listening to several different events we do something more
simple:

Listen for the module to be disposed. When this happens listen for
the reload cycle to be complete. When it is complete we can remove the
old style. This is safe because:

1) We always want to remove the old style. So if a css module is
reloaded we always want to remove the old one.
2) By always removing the old we ensure orphans are removed as well.

The key feature of this refactor is that we no longer listen for the
reload event for the css module. We don't care if that css module is
reloaded or not. We only care that 1) it is disposed 2) the reload cycle
finishes (at some point).